requirements.txt
numpy
opencv-python
pypng